Set new default shader for Apple platforms and IOS 15 support #jira UE-131184 #jira UE-124744 #rb jack.porter #fyi will.damon carl.lloyd #preflight 618e2f249e784835d570d871
[CL 18165990 by axel riffard in 5.0 branch]

+++ Engine/Source/Developer/Apple/MetalShaderFormat/Private/MetalDerivedData.h
@@ -135,7 +135,7 @@ struct FMetalShaderPreprocessed
 class FMetalShaderOutputCooker : public FDerivedDataPluginInterface
 {
 public:
-	FMetalShaderOutputCooker(const FShaderCompilerInput& _Input,FShaderCompilerOutput& Output,const FString& WorkingDirectory, FString PreprocessedShader, FSHAHash GUIDHash, uint8 VersionEnum, uint32 CCFlags, EMetalGPUSemantics Semantics, EMetalTypeBufferMode TypeMode, uint32 MaxUnrollLoops, EShaderFrequency Frequency, bool bDumpDebugInfo, FString Standard, FString MinOSVersion);
+	FMetalShaderOutputCooker(const FShaderCompilerInput& _Input,FShaderCompilerOutput& Output,const FString& WorkingDirectory, FString PreprocessedShader, FSHAHash GUIDHash, uint32 VersionEnum, uint32 CCFlags, EMetalGPUSemantics Semantics, EMetalTypeBufferMode TypeMode, uint32 MaxUnrollLoops, EShaderFrequency Frequency, bool bDumpDebugInfo, FString Standard, FString MinOSVersion);
 	virtual ~FMetalShaderOutputCooker();
 
 #if PLATFORM_MAC || PLATFORM_IOS
